move get_colorbar_and_data_ranges from img_plotting to _utils
